+++ agent/agent-core/src/main/java/org/bithon/agent/core/plugin/PluginInterceptorInstaller.java
@@ -20,7 +20,7 @@ import org.bithon.agent.bootstrap.aop.IBithonObject;
 import org.bithon.agent.bootstrap.loader.JarClassLoader;
 import org.bithon.agent.bootstrap.loader.PluginClassLoaderManager;
 import org.bithon.agent.core.aop.descriptor.Descriptors;
-import org.bithon.agent.core.aop.interceptor.InterceptorInstaller;
+import org.bithon.agent.core.aop.installer.InterceptorInstaller;
 import org.bithon.agent.core.context.AgentContext;
 import org.bithon.component.commons.logging.ILogAdaptor;
 import org.bithon.component.commons.logging.LoggerFactory;
